requirements.txt
beautifulsoup4
requests
python-dateutil
python-dotenv